till section 10 done

+++ src/main/java/com/justin/springboottest/service/StudentAndGradeService.java
@@ -1,9 +1,12 @@
 package com.justin.springboottest.service;
 
 import com.justin.springboottest.models.CollegeStudent;
+import com.justin.springboottest.models.Grade;
+import com.justin.springboottest.models.GradebookCollegeStudent;
 import com.justin.springboottest.models.HistoryGrade;
 import com.justin.springboottest.models.MathGrade;
 import com.justin.springboottest.models.ScienceGrade;
+import com.justin.springboottest.models.StudentGrades;
 import com.justin.springboottest.repository.HistoryGradesDao;
 import com.justin.springboottest.repository.MathGradesDao;
 import com.justin.springboottest.repository.ScienceGradesDao;
@@ -12,6 +15,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -39,6 +44,9 @@ public class StudentAndGradeService {
     @Autowired
     private HistoryGradesDao historyGradesDao;
 
+    @Autowired
+    private StudentGrades studentGrades;
+
     public void createStudent(String firstname, String lastname, String email){
         CollegeStudent student = new CollegeStudent(firstname,lastname,email);
 //        student.setId(0);
@@ -54,7 +62,13 @@ public class StudentAndGradeService {
     }
 
     public void deleteStudent(int id) {
-        studentDao.deleteById(id);
+        boolean studentExists = checkIfStudentExists(id);
+        if (studentExists){
+            studentDao.deleteById(id);
+            mathGradesDao.deleteGradeByStudentId(id);
+            historyGradesDao.deleteGradeByStudentId(id);
+            scienceGradesDao.deleteGradeByStudentId(id);
+        }
     }
 
     public Iterable<CollegeStudent> getGradebook() {
@@ -93,6 +107,70 @@ public class StudentAndGradeService {
         }
 
         return false;
+    }
+
+    public int deleteGrade(int gradeId, String gradeType) {
+        int studentId = 0;
+
+        if (gradeType.equals("math")){
+            Optional<MathGrade> mathGrade = mathGradesDao.findById(gradeId);
+            if(!mathGrade.isPresent()){
+                return studentId;
+            }
+            studentId = mathGrade.get().getStudentId();
+            mathGradesDao.deleteById(gradeId);
+            return studentId;
+        } else if (gradeType.equals("science")){
+            Optional<ScienceGrade> scienceGrade = scienceGradesDao.findById(gradeId);
+            if(!scienceGrade.isPresent()){
+                return studentId;
+            }
+            studentId = scienceGrade.get().getStudentId();
+            scienceGradesDao.deleteById(gradeId);
+            return studentId;
+        } else if (gradeType.equals("history")){
+            Optional<HistoryGrade> historyGrade = historyGradesDao.findById(gradeId);
+            if(!historyGrade.isPresent()){
+                return studentId;
+            }
+            studentId = historyGrade.get().getStudentId();
+            historyGradesDao.deleteById(gradeId);
+            return studentId;
+        }
+
+        return studentId;
+    }
+
+    public GradebookCollegeStudent getStudentInformation(int studentId) {
+
+        boolean studentExists = checkIfStudentExists(studentId);
+        if (!studentExists)
+            return null;
+
+        Optional<CollegeStudent> student = studentDao.findById(studentId);
+        Iterable<MathGrade> mathGrades = mathGradesDao.findGradeByStudentId(studentId);
+        Iterable<ScienceGrade> scienceGrades = scienceGradesDao.findGradeByStudentId(studentId);
+        Iterable<HistoryGrade> historyGrades = historyGradesDao.findGradeByStudentId(studentId);
+
+        List<Grade> mathGradeList = new ArrayList<>();
+        mathGrades.forEach(mathGradeList::add);
+        List<Grade> scienceGradeList = new ArrayList<>();
+        scienceGrades.forEach(scienceGradeList::add);
+        List<Grade> historyGradeList = new ArrayList<>();
+        historyGrades.forEach(historyGradeList::add);
+
+        studentGrades.setMathGradeResults(mathGradeList);
+        studentGrades.setScienceGradeResults(scienceGradeList);
+        studentGrades.setHistoryGradeResults(historyGradeList);
+
+        CollegeStudent collegeStudent = student.get();
+        GradebookCollegeStudent gradebookCollegeStudent = new GradebookCollegeStudent(collegeStudent.getId()
+                , collegeStudent.getFirstname()
+                , collegeStudent.getLastname()
+                , collegeStudent.getEmailAddress()
+                , studentGrades);
+
+        return gradebookCollegeStudent;
     }
 }
